Updated the path of import for "logo.svg"
Path before --> ../../logo.svg (which is giving the error as the path is not correct)
Path after --> ../../assets/logo.svg (correct path)

Please accept my PR and merge the changes as i am a beginner in open source also and want to contribute to this repo and thankyou for making all this stuff. Really helpful. Doing a great job bro.